requirements.txt
Django==2.2.4
django-tastypie==0.14.2
psycopg2==2.8.3
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
